IEEE Coding Challenge
Refactored test wrapper to NamedConsumer
Added stat-tracking
Need to fix test timeout (doesn't kill the thread after timeout)

+++ Miscellaneous/src/ieee_coding_challenge/q1/FibonacciFast.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 
 public class FibonacciFast extends Fibonacci {
 
-    private final ArrayList<Integer> results;
+    private final ArrayList<Long> results;
 
     public FibonacciFast() {
         results = new ArrayList<>();
-        results.add(1);
-        results.add(1);
+        results.add(1L);
+        results.add(1L);
     }
 
     @Override
